# Conflicts:
#	3D_Game/TeamB_Game/game.cpp

+++ 3D_Game/TeamB_Game/meshfield.cpp
@@ -199,7 +199,7 @@ void SetMeshfield(D3DXVECTOR3 pos, D3DXVECTOR3 rot, int textype, int nDiviX,int 
 				for (int nCntX = 0; nCntX <= g_Meshfield[nCnt].nDiviX; nCntX++)
 				{
 					//頂点座標の設定
-					pVtx[0].pos = D3DXVECTOR3(-g_Meshfield[nCnt].nWidth + (g_Meshfield[nCnt].nWidth * nCntX), 0.0f, g_Meshfield[nCnt].nHeight - (g_Meshfield[nCnt].nHeight * nCntZ));
+					pVtx[0].pos = D3DXVECTOR3((float)(-g_Meshfield[nCnt].nWidth + (g_Meshfield[nCnt].nWidth * nCntX)), 0.0f, (float)(g_Meshfield[nCnt].nHeight - (g_Meshfield[nCnt].nHeight * nCntZ)));
 
 					//法線ベクトルの設定
 					pVtx[0].nor = D3DXVECTOR3(0.0f, 1.0f, 0.0f);
@@ -246,7 +246,6 @@ void SetMeshfield(D3DXVECTOR3 pos, D3DXVECTOR3 rot, int textype, int nDiviX,int 
 		}
 	}
 }
-
 //===============================
 // メッシュ床のテクスチャ設定処理
 //===============================
